X-cli-main/pyproject.toml
[tool.poetry]
name = "x-cli"
version = "0.1.0"
description = ""
authors = ["Akiho Kawada <kawadatakeshiho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tweepy = "^4.14.0"
python-dotenv = "^1.0.0"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
